move utils inside orientation
