+++ app/components/PinList.js
@@ -1,6 +1,6 @@
 import React, {Component, View, ListView} from 'react-native';
 import PinListItem from './PinListItem.js';
-import * as geoAction from '../lib/utils';
+import * as geoAction from '../lib/orientation/utils';
 
 export default class PinList extends Component {
